+++ src/test/java/de/alpharogroup/collections/array/ArrayExtensionsTest.java
@@ -25,10 +25,8 @@
 package de.alpharogroup.collections.array;
 
 import static org.testng.AssertJUnit.assertEquals;
-import static org.testng.AssertJUnit.assertTrue;
 
 import java.lang.reflect.InvocationTargetException;
-import java.util.Arrays;
 import java.util.List;
 
 import org.meanbean.test.BeanTestException;
@@ -359,62 +357,6 @@ public class ArrayExtensionsTest
 		expected = false;
 		actual = numbers.isLast(null);
 		assertEquals(expected, actual);
-	}
-
-	/**
-	 * Test for method {@link ArrayExtensions#newArray(Object...)}
-	 */
-	@SuppressWarnings("deprecation")
-	@Test
-	public void testNewArray()
-	{
-		Integer actual;
-		Integer expected;
-		final Integer[] numbers = { 1, 2, 3 };
-		final Integer[] array = ArrayExtensions.newArray(1, 2, 3);
-		for (int i = 0; i < numbers.length; i++)
-		{
-			expected = numbers[i];
-			actual = array[i];
-			assertEquals(expected, actual);
-		}
-
-		final Object[] objects = ArrayExtensions.newArray();
-		expected = 0;
-		actual = objects.length;
-		assertEquals(expected, actual);
-	}
-
-	/**
-	 * Test the method {@link ArrayExtensions#newRangeArray(int, int)}
-	 */
-	@SuppressWarnings("deprecation")
-	@Test
-	public void testNewRangeArray()
-	{
-		Integer[] actual;
-		Integer[] expected;
-		actual = ArrayExtensions.newRangeArray(5, 9);
-		expected = ArrayExtensions.newArray(5, 6, 7, 8, 9);
-		assertTrue(Arrays.deepEquals(actual, expected));
-
-		actual = ArrayExtensions.newRangeArray(1, 49);
-		expected = ArrayExtensions.newArray(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16,
-			17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38,
-			39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49);
-
-		assertTrue(Arrays.deepEquals(actual, expected));
-	}
-
-	/**
-	 * Test the method {@link ArrayExtensions#newRangeArray(int, int)} where end is smaller then
-	 * start
-	 */
-	@SuppressWarnings("deprecation")
-	@Test(expectedExceptions = IllegalArgumentException.class)
-	public void testNewRangeArrayException()
-	{
-		ArrayExtensions.newRangeArray(9, 8);
 	}
 
 	/**